-bişiler

+++ GTBack/DTO/AttrDto.cs
@@ -8,7 +8,7 @@ namespace GTBack.Core.DTO
 {
     public class AttrDto
     {
-        public int placeId { get; set; }
+        public long placeId { get; set; }
         public string Name { get; set; }
         public bool isExist { get; set; }
     }
