+++ GTBack/Entities/ExtensionStrings.cs
@@ -10,7 +10,7 @@ namespace GTBack.Core.Entities
     {
 
 
-        public int placeId { get; set; }
+        public long placeId { get; set; }
 
         public Place Place { get; set; }
